% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author-checklist.R
\name{advise_author}
\alias{advise_author}
\title{Create a data product checklist in a GitHub issue}
\usage{
advise_author(repo = NULL)
}
\arguments{
\item{repo}{GitHub repository of the data product e.g. \code{"maelle/seadog"}}
}
\description{
Create a data product checklist in a GitHub issue
}
